Add AuthorizationService

+++ src/MuscleMemory.Application/Exercies/Commands/DeleteExercise/DeleteExerciseCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using MuscleMemory.Application.Exercies.Commands.CreateExercise;
 using MuscleMemory.Application.Exercies.Commands.UpdateRecord;
@@ -7,12 +8,14 @@ using MuscleMemory.Application.Users;
 using MuscleMemory.Domain.Entities;
 using MuscleMemory.Domain.Exeptions;
 using MuscleMemory.Domain.Repositories;
+using MuscleMemory.Infrastructure.Authorization.Services;
 
 namespace MuscleMemory.Application.Exercies.Commands.DeleteExercise;
 
 public class DeleteExerciseCommandHandler(ILogger<DeleteExerciseCommandHandler> logger,
     IExerciseRepository exerciseRepository,
-    IUserContext userContext) : IRequestHandler<DeleteExerciseCommand>
+    IUserContext userContext,
+    IExerciseAuthorizationService authorizationService) : IRequestHandler<DeleteExerciseCommand>
 {
     public async Task Handle(DeleteExerciseCommand request, CancellationToken cancellationToken)
     {
@@ -23,6 +26,8 @@ public class DeleteExerciseCommandHandler(ILogger<DeleteExerciseCommandHandler> 
         var exercise = await exerciseRepository.GetUserExerciseByIdAsync(request.Id);
 
         if (exercise == null) throw new NotFoundException(nameof(Exercise), request.Id.ToString());
+
+        if (!authorizationService.Authorize(exercise)) throw new ForbidException();
 
         await exerciseRepository.DeleteUserExerciseById(exercise);
     }
